+++ mayan/apps/django_gpg/tests/test_models.py
@@ -18,9 +18,9 @@ from ..models import Key
 
 from .literals import (
     MOCK_SEARCH_KEYS_RESPONSE, TEST_DETACHED_SIGNATURE, TEST_FILE,
-    TEST_KEY_DATA, TEST_KEY_FINGERPRINT, TEST_KEY_PASSPHRASE,
-    TEST_SEARCH_FINGERPRINT, TEST_SEARCH_UID, TEST_SIGNED_FILE,
-    TEST_SIGNED_FILE_CONTENT
+    TEST_KEY_PRIVATE_DATA, TEST_KEY_PRIVATE_FINGERPRINT,
+    TEST_KEY_PRIVATE_PASSPHRASE, TEST_SEARCH_FINGERPRINT, TEST_SEARCH_UID,
+    TEST_SIGNED_FILE, TEST_SIGNED_FILE_CONTENT
 )
 from .mocks import mock_recv_keys
 
@@ -28,9 +28,9 @@ from .mocks import mock_recv_keys
 class KeyTestCase(BaseTestCase):
     def test_key_instance_creation(self):
         # Creating a Key instance is analogous to importing a key
-        key = Key.objects.create(key_data=TEST_KEY_DATA)
+        key = Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
-        self.assertEqual(key.fingerprint, TEST_KEY_FINGERPRINT)
+        self.assertEqual(key.fingerprint, TEST_KEY_PRIVATE_FINGERPRINT)
 
     @mock.patch.object(gnupg.GPG, 'search_keys', autospec=True)
     def test_key_search(self, search_keys):
@@ -69,36 +69,36 @@ class KeyTestCase(BaseTestCase):
         with open(TEST_SIGNED_FILE, mode='rb') as signed_file:
             result = Key.objects.verify_file(signed_file)
 
-        self.assertTrue(result.key_id in TEST_KEY_FINGERPRINT)
+        self.assertTrue(result.key_id in TEST_KEY_PRIVATE_FINGERPRINT)
 
     def test_embedded_verification_with_key(self):
-        Key.objects.create(key_data=TEST_KEY_DATA)
+        Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
         with open(TEST_SIGNED_FILE, mode='rb') as signed_file:
             result = Key.objects.verify_file(signed_file)
 
-        self.assertEqual(result.fingerprint, TEST_KEY_FINGERPRINT)
+        self.assertEqual(result.fingerprint, TEST_KEY_PRIVATE_FINGERPRINT)
 
     def test_embedded_verification_with_correct_fingerprint(self):
-        Key.objects.create(key_data=TEST_KEY_DATA)
+        Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
         with open(TEST_SIGNED_FILE, mode='rb') as signed_file:
             result = Key.objects.verify_file(
-                signed_file, key_fingerprint=TEST_KEY_FINGERPRINT
+                signed_file, key_fingerprint=TEST_KEY_PRIVATE_FINGERPRINT
             )
 
         self.assertTrue(result.valid)
-        self.assertEqual(result.fingerprint, TEST_KEY_FINGERPRINT)
+        self.assertEqual(result.fingerprint, TEST_KEY_PRIVATE_FINGERPRINT)
 
     def test_embedded_verification_with_incorrect_fingerprint(self):
-        Key.objects.create(key_data=TEST_KEY_DATA)
+        Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
         with open(TEST_SIGNED_FILE, mode='rb') as signed_file:
             with self.assertRaises(KeyDoesNotExist):
                 Key.objects.verify_file(signed_file, key_fingerprint='999')
 
     def test_signed_file_decryption(self):
-        Key.objects.create(key_data=TEST_KEY_DATA)
+        Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
         with open(TEST_SIGNED_FILE, mode='rb') as signed_file:
             result = Key.objects.decrypt_file(file_object=signed_file)
@@ -122,10 +122,10 @@ class KeyTestCase(BaseTestCase):
                     file_object=test_file, signature_file=signature_file
                 )
 
-        self.assertTrue(result.key_id in TEST_KEY_FINGERPRINT)
+        self.assertTrue(result.key_id in TEST_KEY_PRIVATE_FINGERPRINT)
 
     def test_detached_verification_with_key(self):
-        Key.objects.create(key_data=TEST_KEY_DATA)
+        Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
         with open(TEST_DETACHED_SIGNATURE, mode='rb') as signature_file:
             with open(TEST_FILE, mode='rb') as test_file:
@@ -134,10 +134,10 @@ class KeyTestCase(BaseTestCase):
                 )
 
         self.assertTrue(result)
-        self.assertEqual(result.fingerprint, TEST_KEY_FINGERPRINT)
+        self.assertEqual(result.fingerprint, TEST_KEY_PRIVATE_FINGERPRINT)
 
     def test_detached_signing_no_passphrase(self):
-        key = Key.objects.create(key_data=TEST_KEY_DATA)
+        key = Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
         with self.assertRaises(NeedPassphrase):
             with open(TEST_FILE, mode='rb') as test_file:
@@ -146,7 +146,7 @@ class KeyTestCase(BaseTestCase):
                 )
 
     def test_detached_signing_bad_passphrase(self):
-        key = Key.objects.create(key_data=TEST_KEY_DATA)
+        key = Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
         with self.assertRaises(PassphraseError):
             with open(TEST_FILE, mode='rb') as test_file:
@@ -156,12 +156,12 @@ class KeyTestCase(BaseTestCase):
                 )
 
     def test_detached_signing_with_passphrase(self):
-        key = Key.objects.create(key_data=TEST_KEY_DATA)
+        key = Key.objects.create(key_data=TEST_KEY_PRIVATE_DATA)
 
         with open(TEST_FILE, mode='rb') as test_file:
             detached_signature = key.sign_file(
                 file_object=test_file, detached=True,
-                passphrase=TEST_KEY_PASSPHRASE
+                passphrase=TEST_KEY_PRIVATE_PASSPHRASE
             )
 
         signature_file = io.BytesIO()
@@ -175,4 +175,4 @@ class KeyTestCase(BaseTestCase):
 
         signature_file.close()
         self.assertTrue(result)
-        self.assertEqual(result.fingerprint, TEST_KEY_FINGERPRINT)
+        self.assertEqual(result.fingerprint, TEST_KEY_PRIVATE_FINGERPRINT)