+++ HISTORY.rst
@@ -110,6 +110,8 @@
   own module.
 - Update label and icon of the document sign form
   Label updated from "Save" to "Sign".
+- Add list, create, detail and edit API views for
+  detached and embedded signatures.
 
 3.2.9 (2019-11-03)
 ==================